added some tests, remove call to previous reset_optimizer
